[IRGen] Expose LinkEntity and ValueWitness publicly.
